Respond to peer changes by resetting without deleting stored data
This isn't strictly correct because starting and stopping sync are
still divorced from setting up SPV. Also unlike enablement and Wifi
sync settings, the service sets peer config rather than SPV.

Before we can fix the above we need to clean up start/stop and sync
more.

+++ app/src/main/java/com/greenaddress/greenbits/spv/SPV.java
@@ -94,10 +94,8 @@ public class SPV {
         new AsyncTask<Object, Object, Object>() {
             @Override
             protected Object doInBackground(Object[] params) {
-                final boolean alreadySyncing = stopSPVSync();
-                setUpSPV();
-                if (alreadySyncing)
-                    startSPVSync();
+                final boolean deleteAllData = false;
+                reset(deleteAllData);
                 return null;
             }
         }.execute();
@@ -617,22 +615,23 @@ public class SPV {
         // FIXME
     }
 
-    public void reset() {
+    public void reset(final boolean deleteAllData) {
         final boolean enabled = mService.isSPVEnabled();
         if (enabled)
             stopSPVSync();
 
-        // Delete all data
-        mService.getSPVChainFile().delete();
+        if (deleteAllData) {
+            mService.getSPVChainFile().delete();
 
-        try {
-            mService.cfgInEdit(SPENDABLE).clear().commit();
-            mService.cfgInEdit(VERIFIED).clear().commit();
-        } catch (final NullPointerException e) {
-            // ignore
+            try {
+                mService.cfgInEdit(SPENDABLE).clear().commit();
+                mService.cfgInEdit(VERIFIED).clear().commit();
+            } catch (final NullPointerException e) {
+                // ignore
+            }
+
+            resetUnspent();
         }
-
-        resetUnspent();
 
         if (enabled) {
             // Restart SPV
